[CodeGen] Refactor transform dialect patterns to Transforms/Patterns.
The patterns can be reused by any other passes. It makes the gaps
between default codegen transform scripts smaller. We also found
FoldFillIntoPad pattern useful in sdxl sprint.
