+++ compiler/src/iree/compiler/Codegen/Transforms/BUILD.bazel
@@ -17,9 +17,11 @@ iree_compiler_cc_library(
     srcs = [
         "AffineMinDistributedSCFCanonicalization.cpp",
         "RemoveSingleIterationLoop.cpp",
+        "Patterns.cpp",
         "Transforms.cpp",
     ],
     hdrs = [
+        "Patterns.h",
         "Transforms.h",
     ],
     deps = [
@@ -42,6 +44,7 @@ iree_compiler_cc_library(
         "@llvm-project//mlir:Pass",
         "@llvm-project//mlir:SCFTransforms",
         "@llvm-project//mlir:Support",
+        "@llvm-project//mlir:TensorDialect",
         "@llvm-project//mlir:TransformUtils",
         "@llvm-project//mlir:Transforms",
         "@llvm-project//mlir:ValueBoundsOpInterface",
